add new files to spread out load
